requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
Faker==25.2.0
idna==3.7
iniconfig==2.0.0
invoke==2.2.0
packaging==24.0
paramiko==3.4.0
pluggy==1.5.0
pycparser==2.22
PyNaCl==1.5.0
pytest==8.2.1
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
urllib3==2.2.1
wrapt==1.16.0
